Create Configure-GameMakerPatch.bat
This is a temporary workaround; it is unsupported by myself or anyone at YoYoGames, and was primarily made for personal use as my installation stopped working. This comes with no warranty whatsoever and I accept zero liability of any kind of damages - review the script yourself and only use it if you deem it both safe and appropriate.