Add files via upload
Temporary workaround for server issues. Provided with zero warranty, I accept no liability for anything which may arise from your use of this script - it is a short script and the XML can be easily expanded; review it yourself and ensure it is safe and appropriate for your use case.